+++ src/main/java/br/com/nfast/api/model/adm/RateioDespesaId.java
@@ -15,7 +15,7 @@ public class RateioDespesaId implements Serializable {
     @JoinColumn(name = "seq_despesa")
     private Despesa despesa;
     @Column(name = "cod_centro_custo", insertable = false, updatable = false)
-    private Integer codCentroCusto;
+    private Long codCentroCusto;
 
     public RateioDespesaId() {
 
@@ -29,11 +29,11 @@ public class RateioDespesaId implements Serializable {
         this.despesa = despesa;
     }
 
-    public Integer getCodCentroCusto() {
+    public Long getCodCentroCusto() {
         return codCentroCusto;
     }
 
-    public void setCodCentroCusto(Integer codCentroCusto) {
+    public void setCodCentroCusto(Long codCentroCusto) {
         this.codCentroCusto = codCentroCusto;
     }
