+++ src/main/java/br/com/nfast/api/model/estoque/PrecoProduto.java
@@ -12,9 +12,9 @@ public class PrecoProduto {
     @JsonIgnore
     private String id;
     @Column(name = "cod_item")
-    private Integer codItem;
+    private Long codItem;
     @Column(name = "cod_empresa")
-    private Integer codEmpresa;
+    private Long codEmpresa;
     @Column(name = "val_preco_venda")
     private Double valPrecoVenda = 0.0;
     @Column(name = "per_margem")
@@ -30,19 +30,19 @@ public class PrecoProduto {
         this.id = id;
     }
 
-    public Integer getCodItem() {
+    public Long getCodItem() {
         return codItem;
     }
 
-    public void setCodItem(Integer codItem) {
+    public void setCodItem(Long codItem) {
         this.codItem = codItem;
     }
 
-    public Integer getCodEmpresa() {
+    public Long getCodEmpresa() {
         return codEmpresa;
     }
 
-    public void setCodEmpresa(Integer codEmpresa) {
+    public void setCodEmpresa(Long codEmpresa) {
         this.codEmpresa = codEmpresa;
     }
