+++ src/main/java/br/com/nfast/api/model/estoque/ProdutoNcm.java
@@ -10,15 +10,15 @@ import javax.persistence.Table;
 public class ProdutoNcm {
     @Id
     @Column(name = "cod_item")
-    private Integer codItem;
+    private Long codItem;
     @Column(name = "seq_ncm")
     private Integer seqNcm;
 
-    public Integer getCodItem() {
+    public Long getCodItem() {
         return codItem;
     }
 
-    public void setCodItem(Integer codItem) {
+    public void setCodItem(Long codItem) {
         this.codItem = codItem;
     }
